refactor: Move env to utils
