+++ cmd/root.go
@@ -3,13 +3,15 @@ package cmd
 import (
 	"fmt"
 	"log"
+
+	"github.com/caffeine-addictt/nexis/cmd/utils"
 )
 
 func init() {
-	loadEnvironment()
+	utils.LoadEnvironment()
 }
 
 func Execute() {
-	server := NewAPIServer(fmt.Sprintf(":%d", Environment.Port))
+	server := NewAPIServer(fmt.Sprintf(":%d", utils.Environment.Port))
 	log.Fatalf("Server crashed: %s", server.Run())
 }
